% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_summarize.R
\name{compute_by_group}
\alias{compute_by_group}
\title{compute by group}
\usage{
compute_by_group(data, expr, by)
}
\arguments{
\item{data}{data}

\item{expr}{expression}

\item{by}{variables to group by}
}
\description{
\code{compute_by_group()} is meant to be used through the \code{{expr} ~ by} syntax
in a call to the nakedpipe, and might be encountered when using the debugging pipe.
}
