% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_functionals.R
\name{\%lapply.\%}
\alias{\%lapply.\%}
\alias{\%sapply.\%}
\title{nakedpipe functionals}
\usage{
X \%lapply.\% expr

X \%sapply.\% expr
}
\arguments{
\item{X}{list or vector of element to apply the sequence on}

\item{expr}{call or sequence of calls wrapped inside \code{{}}}
}
\description{
These combine functionals and \verb{\%.\%}
}
