% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_snippets.R
\name{setup_nakedpipe_snippets}
\alias{setup_nakedpipe_snippets}
\title{setup nakedpipe snippets}
\usage{
setup_nakedpipe_snippets()
}
\description{
Prints snippet definitions and attempts to open in the editor the
snippet file to modify.
}
