% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nakedpipe.R
\name{\%.\%}
\alias{\%.\%}
\alias{\%..\%}
\title{Naked pipe}
\usage{
x \%.\% expr

x \%..\% expr
}
\arguments{
\item{x}{object to pipe in}

\item{expr}{sequence of calls wrapped inside `{}`}
}
\value{

}
\description{
Pipe into a sequence of calls, without repeating the pipe symbol
}
\details{
This is inspired by Stefan Bache and Hadley Wickham's *magrittr* pipe and
behaves consistently, though it doesn't use its code.

If you want to be even faster, use `%..%` and use
explicit dots (though keep in mind these are micro seconds!)
}
\examples{
# Pipe into a sequence of calls using `\%.\%`
cars \%.\% {
  head(2)
  transform(time = dist/speed)
}

# `\%..\%` is faster at the cost of using explicit dots
cars \%..\% {
  head(.,2)
  transform(.,time = dist/speed)
}

# the naked pipe plays well with left to right assignment
cars \%.\% {
  head(2)
  transform(time = dist/speed)
} -> res
res
}
