% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_snippets.R
\name{setup_nakedpipe_snippets}
\alias{setup_nakedpipe_snippets}
\title{setup nakedpipe snippets}
\usage{
setup_nakedpipe_snippets()
}
\description{
Places snippet definitions in the pastebin and opens in the editor the
snippet file to modify.
}
